requirements.txt
pyautogen==0.9.0
selenium==4.33.0
webdriver-manager==4.0.2
openai==1.82.0
beautifulsoup4>=4.13.0
python-dotenv>=1.0.0
